wip(simulator): Removes local transaction option, adds target to config
