+++ simulator/src/delegate.rs
@@ -3,33 +3,29 @@ use mugraph_core::{crypto, error::Error, types::*};
 use mugraph_node::database::Database;
 use rand::prelude::*;
 use rand_chacha::ChaCha20Rng;
+use reqwest::blocking::Client;
 use tracing::info;
 
-use crate::node::{Node, NodeTarget};
+// use crate::node::{Node, NodeTarget};
 
 #[derive(Debug)]
 pub struct Delegate {
     pub rng: ChaCha20Rng,
     pub db: Database,
     pub keypair: Keypair,
-    pub node: Node,
+    client: Client,
+    target: String,
 }
 
 impl Delegate {
-    pub fn new<R: Rng + CryptoRng>(rng: &mut R, keypair: Keypair, target: Option<String>) -> Result<Self, Error> {
+    pub fn new<R: Rng + CryptoRng>(rng: &mut R, keypair: Keypair, target: String) -> Result<Self, Error> {
         let mut rng = ChaCha20Rng::seed_from_u64(rng.gen());
 
         info!(public_key = %keypair.public_key, "Starting delegate");
         let db = Database::setup_test(&mut rng, None)?;
 
-        let node_target = match target {
-          Some(endpoint) => NodeTarget::Remote(endpoint),
-          None => NodeTarget::Local
-        };
-
-        let node = Node::new(node_target).unwrap();
-
-        Ok(Self { db, rng, keypair, node })
+        let client = Client::new();
+        Ok(Self { db, rng, keypair, client, target })
     }
 
     #[tracing::instrument(skip_all)]
@@ -53,6 +49,32 @@ impl Delegate {
     #[inline(always)]
     #[tracing::instrument(skip_all)]
     pub fn recv_transaction_v0(&mut self, tx: &Transaction) -> Result<V0Response, Error> {
-      self.node.execute_transaction_v0(tx, self.keypair, &mut self.db)
+      let target_endpoint = format!("{}/v0/rpc", self.target);
+      let request = Request::V0(V0Request::Transaction(tx.clone()));
+
+      let response = self
+        .client
+        .post(&target_endpoint)
+        .json(&request)
+        .send()
+        .map_err(|err| Error::ServerError {
+          reason: err.to_string(),
+        })?;
+
+      let response_text = response.text().map_err(|err| Error::ServerError {
+        reason: err.to_string(),
+      })?;
+      let v0_response: V0Response = serde_json::from_str(&response_text).map_err(|_| {
+        if response_text.contains("Atom has already been spent") {
+          return Error::AlreadySpent {
+            signature: tx.signatures[0],
+          };
+        }
+        Error::ServerError {
+          reason: response_text,
+        }
+      })?;
+
+      Ok(v0_response)
     }
 }
