+++ simulator/src/main.rs
@@ -43,7 +43,8 @@ fn main() -> Result<()> {
         let ir = is_running.clone();
         let ip = is_preparing.clone();
         let seed: u64 = rng.gen();
-
+        let node_target = config.node_endpoint.clone();
+        
         thread::spawn(move || {
             core_affinity::set_for_current(core);
 
@@ -51,8 +52,7 @@ fn main() -> Result<()> {
 
             let mut rng = ChaCha20Rng::seed_from_u64(seed);
 
-            let remote_target = Some("http://localhost:9999".to_string());
-            let delegate = Delegate::new(&mut rng, keypair, remote_target)?;
+            let delegate = Delegate::new(&mut rng, keypair, node_target)?;
             let mut sim = Simulation::new(&mut rng, core.id as u32, delegate)?;
 
             while ip.load(Ordering::Relaxed) {
